@@ -1,15 +1,14 @@
 import {Component, input} from '@angular/core';
 import {ProjectMetadata} from "../../../dto/projects/ProjectMetadata";
-import {DatePipe, NgIf} from "@angular/common";
+import { DatePipe } from "@angular/common";
 import {RouterLink} from "@angular/router";
 
 @Component({
     selector: 'project-card',
     imports: [
-        NgIf,
-        RouterLink,
-        DatePipe
-    ],
+    RouterLink,
+    DatePipe
+],
     templateUrl: './project-card.component.html',
     standalone: true,
     styleUrl: './project-card.component.css'
