@@ -3,7 +3,7 @@ import {JollyService} from "../../jolly.service";
 import {GameSessionGameComponent} from "../../../all/components/game-session-game.component";
 import {GameState} from "../../dto/game/GameState";
 import {AuthService} from "../../../../main/services/auth.service";
-import {NgIf} from "@angular/common";
+
 import {Router} from "@angular/router";
 import {RoundPopupComponent} from "../popups/round-popup/round-popup.component";
 import {RoundResult} from "../../dto/RoundResult";
@@ -15,11 +15,10 @@ import {WithFile} from "../../../../main/dto/all/WithFile";
   selector: 'jolly-game',
   imports: [
     GameSessionGameComponent,
-    NgIf,
     RoundPopupComponent,
     ReactiveFormsModule,
     JollyRoundTableComponent
-  ],
+],
   templateUrl: './game.component.html',
   standalone: true,
   styleUrl: './game.component.css'
