@@ -8,7 +8,7 @@ import {GameConfig} from "../../dto/game/GameConfig";
 import {GameSessionConfigComponent} from "../../../all/components/game-session-config.component";
 import {PlayerSelectComponent} from "../../../../main/components/all/player-select/player-select.component";
 import {SliderCheckboxComponent} from "../../../../main/components/all/slider-checkbox/slider-checkbox.component";
-import {NgForOf} from "@angular/common";
+
 import {GameMode, getGameModeName} from "../../../catan/dto/enums/GameMode";
 
 @Component({
@@ -17,9 +17,8 @@ import {GameMode, getGameModeName} from "../../../catan/dto/enums/GameMode";
     GameSessionConfigComponent,
     PlayerSelectComponent,
     ReactiveFormsModule,
-    SliderCheckboxComponent,
-    NgForOf
-  ],
+    SliderCheckboxComponent
+],
   templateUrl: './config.component.html',
   styleUrl: './config.component.css'
 })
